+++ webots_ros2_driver/src/Driver.cpp
@@ -18,6 +18,8 @@
 #include <iostream>
 #include <memory>
 #include <rclcpp/rclcpp.hpp>
+#include <webots_ros2_driver/PluginInterface.hpp>
+#include <webots_ros2_driver/PythonPlugin.hpp>
 #include <webots_ros2_driver/WebotsNode.hpp>
 
 int main(int argc, char **argv) {
@@ -67,7 +69,21 @@ int main(int argc, char **argv) {
       break;
     rclcpp::spin_some(node);
   }
-  wb_robot_cleanup();
+  // Check if the plugin is actually an instance of PythonPlugin
+  bool isPython = false;
+  for (std::shared_ptr<webots_ros2_driver::PluginInterface> plugin : node->mPlugins) {
+    std::shared_ptr<webots_ros2_driver::PythonPlugin> pythonPlugin =
+      std::dynamic_pointer_cast<webots_ros2_driver::PythonPlugin>(plugin);
+    if (pythonPlugin) {
+      pythonPlugin->stop();  // stop only for python plugins
+      isPython = true;
+      break;
+    }
+  }
+  if (!isPython) {
+    wb_robot_cleanup();
+  }
+  RCLCPP_INFO(node->get_logger(), "Controller successfully disconnected from robot in Webots simulation.");
   rclcpp::shutdown();
   return 0;
 }