+++ webots_ros2_driver/CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog for package webots_ros2_driver
 ------------------
 * Added component remapping parameter to rename PROTO components.
 * Added deprecation message when declaring driver node in launch file.
+* Added animation_{start,stop}_recording services to Ros2Supervisor node.
+* Added /Ros2Supervisor namespace to Ros2Supervisor node.
+* Fixed Python plugin termination on SIGINT call or simulation ends.
 * Fixed RangeFinder activation to also check for point cloud subscriptions.
 
 2023.1.0 (2023-06-29)